+++ ocdskit/cli/commands/schema_strict.py
@@ -3,6 +3,7 @@ import sys
 from copy import deepcopy
 
 from ocdskit.cli.commands.base import BaseCommand
+from ocdskit.schema import add_validation_properties
 from ocdskit.util import json_dump
 
 
@@ -19,35 +20,11 @@ class Command(BaseCommand):
                           help='check the file for missing properties without modifying the file')
 
     def handle(self):
-        def recurse(data):
-            if isinstance(data, list):
-                for item in data:
-                    recurse(item)
-            elif isinstance(data, dict):
-                if 'type' in data:
-                    if (
-                        'string' in data['type']
-                        and 'enum' not in data
-                        and 'format' not in data
-                        and 'pattern' not in data
-                    ):
-                        data.setdefault('minLength', 1)
-                    if 'array' in data['type']:
-                        data.setdefault('minItems', 1)
-                        if not self.args.no_unique_items:
-                            data.setdefault('uniqueItems', True)
-
-                    if 'object' in data['type']:
-                        data.setdefault('minProperties', 1)
-
-                for value in data.values():
-                    recurse(value)
-
         with open(self.args.file) as f:
             schema = json.load(f)
 
         original = deepcopy(schema)
-        recurse(schema)
+        add_validation_properties(schema, unique_items=not self.args.no_unique_items)
 
         if self.args.check:
             if schema != original: